callHistory/create-call-history/requirements.txt
fdk>=0.1.51
oci
PyJWT
python-dotenv